+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- added: (ALGO) Handle broadcast errors to recipients with unactivated assets
+- added: (ALGO) Throw if recipient did not activate the token being sent in `signTx` 
 
 ## 4.38.0 (2025-03-03)